FLANN skel in, changed ANN distances to floats from doubles
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13070 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: c4ab468388449c2c6032e272daf05cbad2c2841e

+++ mapping/plug_onbase_detector/src/plug_onbase_detector.cpp
@@ -318,7 +318,7 @@ class PlugOnBaseDetector
       processed.resize (indices.size (), false);
 
       vector<int> nn_indices;
-      vector<double> nn_distances;
+      vector<float> nn_distances;
       // Process all points in the indices vector
       for (unsigned int i = 0; i < indices.size (); i++)
       {
